+++ app/(navigatable)/(orgs)/orgs/[id]/page.tsx
@@ -52,14 +52,14 @@ const OrgIndex = ({ params: { id } }: { params: { id: string } }) => {
     return null
   }
 
-  // Once we have loaded our data set the active billing org to it
+  // Once we have loaded our data set the active billing context to it
   setActiveBillingOrg(org)
 
   return (
     <div className="flex-1 flex-col gap-4 p-10 text-2xl font-bold">
       <RenderableResourceContent>
         <div className="flex flex-col items-center">
-          <h3>Current User Context</h3>
+          <h3>Current Billing Context</h3>
           <div className="w-1/2 border-t-2 border-blue-600 my-2"></div>
           <p className="text-lg">{org.name}</p>
         </div>