+++ app/(navigatable)/(projects)/projects/project-dashboard.tsx
@@ -17,6 +17,10 @@ const ProjectDashboard = ({
 }: ProjectDashboardProps) => {
   return (
     <div>
+      <div className="text-center mb-2">Projects</div>
+      <div className="flex justify-center w-full">
+        <div className="w-1/2 border-t-2 border-blue-600 mb-5"></div>
+      </div>
       <div className="mb-10">
         <ProjectCreateTile />
       </div>